+++ modules/frontend/controllers/PostController.php
@@ -71,19 +71,4 @@ class PostController extends BaseController
         ]);
     }
 
-    public function actionRandom() {
-        $entries = Entry::find()
-            ->section('post')
-            ->limit(3)
-            ->orderBy('rand()')
-            ->all();
-
-        return $this->inertia('Site/Index', [
-            'randomPosts' => array_map(fn($entry) => [
-                'id' => $entry->id,
-                'title' => $entry->title,
-                'url' => $entry->inertiaUrl
-            ], $entries)
-        ]);
-    }
 }