+++ modules/frontend/controllers/SiteController.php
@@ -2,6 +2,7 @@
 
 namespace modules\frontend\controllers;
 
+use craft\elements\Entry;
 use craft\elements\GlobalSet;
 use modules\frontend\helpers\HtmlHelper;
 
@@ -11,6 +12,18 @@ class SiteController extends BaseController
     {
         $siteInfo = GlobalSet::find()->handle('siteInfo')->one();
 
+        if ($this->only == 'randomPosts') {
+            $entries = Entry::find()->section('post')->limit(6)->orderBy('rand()')->all();
+
+            return $this->inertia('Site/Index', [
+               'randomPosts' =>  array_map(fn($entry) => [
+                   'id' => $entry->id,
+                   'title' => $entry->title,
+                   'url' => $entry->inertiaUrl,
+               ], $entries),
+                'notice' => 'This is an example for partially refreshing the page.'
+            ]);
+        }
 
         return $this->inertia('Site/Index', [
             'title' => $siteInfo->siteIntoTitle,