+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This is the **first final version** of this Proof of Concept.
 
 * Minor layout improvements
 * Moved Site Nav to custom config
+* Stop running Craft in headless mode
 
 ### Fixed
 
@@ -26,6 +27,8 @@ This is the **first final version** of this Proof of Concept.
 ### Internal
 
 * ImageTag and Blocks vue components
+* Cleanup controllers / custom settings
+* User partial data loading when staying in the same component
 
 ## 1.4.0 2022-06-24